fix: remove broken tests and rename styles file
