+++ tasklist/client/src/modules/bpmn-js/BpmnJS.ts
@@ -9,7 +9,7 @@
 import NavigatedViewer, {BpmnElement} from 'bpmn-js/lib/NavigatedViewer';
 // @ts-expect-error Could not find a declaration file for module '@bpmn-io/element-templates-icons-renderer'
 import ElementTemplatesIconsRenderer from '@bpmn-io/element-template-icon-renderer';
-import {bpmnRendererColors} from './styled';
+import {bpmnRendererColors} from './styles';
 
 type RenderOptions = {
   container: HTMLElement;
